test/keysym: fix comparison of integers of different signs
```
../test/keysym.c:320:9: warning: comparison of integers of different signs: 'int' and 'unsigned long' [-Wsign-compare]
  320 |         U16_APPEND(cp_string, offset, ARRAY_SIZE(cp_string), cp, isError);
      |         ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
/usr/include/unicode/utf16.h:396:47: note: expanded from macro 'U16_APPEND'
  396 |     } else if((uint32_t)(c)<=0x10ffff && (i)+1<(capacity)) { \
      |                                          ~~~~~^ ~~~~~~~~
```

+++ test/keysym.c
@@ -317,7 +317,7 @@ test_icu_case_mappings(xkb_keysym_t ks)
         UBool isError = false;
         int32_t offset = 0;
         /* Convert code point to UTF-16 string */
-        U16_APPEND(cp_string, offset, ARRAY_SIZE(cp_string), cp, isError);
+        U16_APPEND(cp_string, offset, (int32_t) ARRAY_SIZE(cp_string), cp, isError);
         UErrorCode pErrorCode = U_ZERO_ERROR;
         /* Unicode full upper mapping */
         int32_t length = u_strToUpper(cp_expected_string,
